add DebugODisplay for print statement or alternative, Improved Code, adding i64 and f64

+++ examples/f64_to_str.rs
@@ -1,32 +1,32 @@
 use embedded_cotoutf8::{COtoUTF8, DebugODisplay};
 
 fn main() {
-    let num: f32 = 1.0005;
+    let num: f64 = 1.0005;
     let binding = num.coto_utf8();
     let result = core::str::from_utf8(&binding).unwrap();
     println!("{}", result);
 
-    let num: f32 = -123.0005;
+    let num: f64 = -123.0005;
     let binding = num.coto_utf8();
     let result = core::str::from_utf8(&binding).unwrap();
     println!("{}", result);
 
-    let num: f32 = 1524.001;
+    let num: f64 = 1524.001;
     let binding = num.coto_utf8();
     let result = core::str::from_utf8(&binding).unwrap();
     println!("{}", result);
 
-    let num: f32 = 44245.12;
+    let num: f64 = 44245.12;
     let binding = num.coto_utf8();
     let result = core::str::from_utf8(&binding).unwrap();
     println!("{}", result);
 
-    let num: f32 = 442400005.001;
+    let num: f64 = 442400005.001;
     let binding = num.coto_utf8();
     let result = core::str::from_utf8(&binding).unwrap();
     println!("{}", result);
 
     // Alternative use this
-    let num = DebugODisplay(44245.12f32);
+    let num = DebugODisplay(44245.12f64);
     println!("{}", num)
 }
